Overhauled item physics
-Every item now had angular movement with speed decay
-Items calculate and move towards the player if close, and move faster
as they get closer
-All items have gravity by default, except for souls
-Gravity can be disabled on specific items
-Gremlin infestation resolved

+++ master/src/Graphics/ItemGraphics.java
@@ -31,8 +31,8 @@ public class ItemGraphics{
 
 	public void drawItem(Item item, Graphics g, Camera camera) throws Exception{
 		Image itemImage = item.getImage().getScaledInstance(item.getWidth(), item.getHeight(), 0);
-		int relativeX = item.getX() - camera.getCamX();
-		int relativeY = (item.getY() - item.getHeight()) - camera.getCamY();
+		int relativeX = (int) item.getX() - camera.getCamX();
+		int relativeY = ((int) item.getY() - item.getHeight()) - camera.getCamY();
 		g.drawImage(itemImage, relativeX, relativeY, panel);
 		//System.out.println("Drawing item");
 		//System.out.println("Drew frame: " + item.getCurFrame());
